+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * add animation interface [#611](https://github.com/jcubic/jquery.terminal/issues/611)
 * use dash `-` as input file to read from STDIN in `from-ansi` executable
 * improve auth error message
+* add "enter" typing animation with prefix prompt
 ### Bugfix
 * fix rest of the extension files in NodeJS (regression since 2.22.0)
 * fix inheritance of attributes (including styles) and classes in nested formatting